+++ workspace/acceptance/global_init_script_test.go
@@ -1,0 +1,20 @@
+package acceptance
+
+import (
+	"testing"
+
+	"github.com/databrickslabs/databricks-terraform/internal/acceptance"
+)
+
+func TestAccGlobalInitScriptResource_Create(t *testing.T) {
+	acceptance.Test(t, []acceptance.Step{
+		{
+			Template: `
+			resource "databricks_global_init_script" "this" {
+				name = "init-{var.RANDOM}"
+				enabled = false
+				content_base64 = "ZWNobyBoZWxsbw=="
+			}`,
+		},
+	})
+}